requirements.txt
discord.py
config
python-dotenv